backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
python-dotenv>=1.0.0
openai>=1.0.0
tavily-python>=0.3.0
apscheduler>=3.10.0
requests>=2.31.0
pillow>=10.0.0
pytrends>=4.9.0
cryptography>=42.0.0
pydantic[email]>=2.0.0
cloudinary>=1.36.0
